Rename ssh-ident wrapper.
The documentation suggests aliasing ssh to ssh-ident, which is great for
local keys (using it on a virtualbox instance works fine.)

However, when I'm using a remote SSH-agent- e.g. SSH'd in to a VM, using
a security key for my keys- ssh-ident tries to load the keys on the
remote host, which aren't authorized.

I suspect there may be something I can do with an ssh-ident file, but
I'm not immediately seeing something like "use an existing ssh-agent if
one is found."