Add ws command (workspace)
I've found a similar command handy at work: keep a tmux session open for
a particular bug / in a particular repository / for a particular
purpose. The 'r' jump at the beginning is to get new terminal sessions
to land in the same repository.

Doesn't do much for the associated browser windows, but meh.