Some fixes to r.
I've never used `r r`, probably because it's not very memorable. Make it
`r ls instead`.

I was getting errors when trying to `r tcell`, with
github.com/gdamore/tcell definitely in the Go workspace. The shell was
complaining about the cceckman/tcell not existing - which was accurate -
but that terminated r(). `set nonomatch` fixes this - making it `ls`'s
failure rather than the shell's if the glob doesn't match anything.