Use directories in .vim for swap and undo
Tools like `ibazel` watch the local directory for changes, and take some
action when the change happens. This doesn't play well with swapfiles,
which can trigger lots of spurious builds. (It's also a bit frustrating
to see in e.g. `git status`, though I've mostly managed to mask that off
at this point.)

Make Vim put persistent undo files and swap files in a temp directory in
$HOME. I'm not expecting this to take much room, but also update
`unclean` to cover it.